client: move types to corresponding component directories
