+++ client/src/apis/config.service.ts
@@ -1,16 +1,13 @@
-import type { Config } from "../components/types.ts";
-import type { ApiResponse } from "./types.ts";
-import { apiClient } from "./client.ts";
+import type {Config} from "../components/ConfigPage/types.ts";
+import type {ApiResponse} from "./types.ts";
+import {apiClient} from "./client.ts";
 
 export const configService = {
     get: () => {
-        return apiClient<ApiResponse>(`/configs`, { method: "GET" });
-    },
-    update: (config: Config) => {
-        console.log(config);
+        return apiClient<ApiResponse>(`/configs`, {method: "GET"});
+    }, update: (config: Config) => {
         return apiClient<ApiResponse>(`/configs`, {
-            method: "POST",
-            body: JSON.stringify(config),
+            method: "POST", body: JSON.stringify(config),
         });
     },
 };
