+++ client/src/components/DevicePage/StatusPill.tsx
@@ -1,4 +1,4 @@
-import type {DeviceStatus} from "../types.ts";
+import type {DeviceStatus} from "./types.ts";
 import React from "react";
 
 export function StatusPill({status}: { status: DeviceStatus }) {
